requirements.txt
fastapi
uvicorn 
motor
pydantic
python-jose
passlib[bcrypt]
python-dotenv
pydantic[email]